requirements.txt
MetaTrader5
pandas
python-dotenv
requests